🚚 Move frequency arbitraries into /arbitrary
