+++ src/arbitrary/_internals/FrequencyArbitrary.ts
@@ -1,57 +1,22 @@
 import { Random } from '../../random/generator/Random';
 import { Stream } from '../../stream/Stream';
-import { Arbitrary } from './definition/Arbitrary';
-import { convertFromNext, convertToNext } from './definition/Converters';
-import { NextArbitrary } from './definition/NextArbitrary';
-import { NextValue } from './definition/NextValue';
+import { Arbitrary } from '../../check/arbitrary/definition/Arbitrary';
+import { convertFromNext, convertToNext } from '../../check/arbitrary/definition/Converters';
+import { NextArbitrary } from '../../check/arbitrary/definition/NextArbitrary';
+import { NextValue } from '../../check/arbitrary/definition/NextValue';
 import { DepthContext, getDepthContextFor } from './helpers/DepthContext';
-
-/**
- * Conjonction of a weight and an arbitrary used by {@link frequency}
- * in order to generate values
- *
- * @remarks Since 1.18.0
- * @public
- */
-export interface WeightedArbitrary<T> {
-  /**
-   * Weight to be applied when selecting which arbitrary should be used
-   * @remarks Since 0.0.7
-   */
-  weight: number;
-  /**
-   * Instance of Arbitrary
-   * @remarks Since 0.0.7
-   */
-  arbitrary: Arbitrary<T>;
-}
-
-/** @internal */
-interface WeightedNextArbitrary<T> {
-  weight: number;
-  arbitrary: NextArbitrary<T>;
-}
-
-/** @internal */
-type FrequencyArbitraryContext<T> = {
-  selectedIndex: number;
-  originalBias: number | undefined;
-  originalContext: unknown;
-  clonedMrngForFallbackFirst: Random | null;
-  cachedGeneratedForFirst?: NextValue<T>;
-};
 
 /** @internal */
 export class FrequencyArbitrary<T> extends NextArbitrary<T> {
-  readonly summedWarbs: WeightedNextArbitrary<T>[];
+  readonly summedWarbs: _WeightedNextArbitrary<T>[];
   readonly totalWeight: number;
 
-  static from<T>(warbs: WeightedArbitrary<T>[], constraints: FrequencyContraints, label: string): Arbitrary<T> {
+  static from<T>(warbs: _WeightedArbitrary<T>[], constraints: _FrequencyContraints, label: string): Arbitrary<T> {
     if (warbs.length === 0) {
       throw new Error(`${label} expects at least one weigthed arbitrary`);
     }
     let totalWeight = 0;
-    const warbsNext: WeightedNextArbitrary<T>[] = [];
+    const warbsNext: _WeightedNextArbitrary<T>[] = [];
     for (let idx = 0; idx !== warbs.length; ++idx) {
       const currentArbitrary = warbs[idx].arbitrary;
       if (currentArbitrary === undefined) {
@@ -76,8 +41,8 @@ export class FrequencyArbitrary<T> extends NextArbitrary<T> {
   }
 
   private constructor(
-    readonly warbs: WeightedNextArbitrary<T>[],
-    readonly constraints: FrequencyContraints,
+    readonly warbs: _WeightedNextArbitrary<T>[],
+    readonly constraints: _FrequencyContraints,
     readonly context: DepthContext
   ) {
     super();
@@ -110,7 +75,7 @@ export class FrequencyArbitrary<T> extends NextArbitrary<T> {
 
   shrink(value: T, context?: unknown): Stream<NextValue<T>> {
     if (context !== undefined) {
-      const safeContext = context as FrequencyArbitraryContext<T>;
+      const safeContext = context as _FrequencyArbitraryContext<T>;
       const selectedIndex = safeContext.selectedIndex;
       const originalBias = safeContext.originalBias;
       const originalArbitrary = this.summedWarbs[selectedIndex].arbitrary;
@@ -165,7 +130,7 @@ export class FrequencyArbitrary<T> extends NextArbitrary<T> {
     clonedMrngForFallbackFirst: Random | null,
     biasFactor: number | undefined
   ): NextValue<T> {
-    const context: FrequencyArbitraryContext<T> = {
+    const context: _FrequencyArbitraryContext<T> = {
       selectedIndex: idx,
       originalBias: biasFactor,
       originalContext: value.context,
@@ -210,113 +175,31 @@ export class FrequencyArbitrary<T> extends NextArbitrary<T> {
   }
 }
 
-/**
- * Infer the type of the Arbitrary produced by {@link frequency}
- * given the type of the source arbitraries
- *
- * @remarks Since 2.2.0
- * @public
- */
-export type FrequencyValue<Ts extends WeightedArbitrary<unknown>[]> = {
-  [K in keyof Ts]: Ts[K] extends WeightedArbitrary<infer U> ? U : never;
-}[number];
+/** @internal */
+interface _WeightedArbitrary<T> {
+  weight: number;
+  arbitrary: Arbitrary<T>;
+}
 
-/**
- * Constraints to be applied on {@link frequency}
- * @remarks Since 2.14.0
- * @public
- */
-export type FrequencyContraints = {
-  /**
-   * When set to true, the shrinker of frequency will try to check if the first arbitrary
-   * could have been used to discover an issue. It allows to shrink trees.
-   *
-   * Warning: First arbitrary must be the one resulting in the smallest structures
-   * for usages in deep tree-like structures.
-   *
-   * Warning: First arbitrary will not be used if its weight is set to zero.
-   *
-   * @remarks Since 2.14.0
-   */
+/** @internal */
+interface _WeightedNextArbitrary<T> {
+  weight: number;
+  arbitrary: NextArbitrary<T>;
+}
+
+/** @internal */
+type _FrequencyContraints = {
   withCrossShrink?: boolean;
-  /**
-   * While going deeper and deeper within a recursive structure (see {@link letrec}),
-   * this factor will be used to increase the probability to generate instances
-   * of the first passed arbitrary.
-   *
-   * Example of values: 0.1 (small impact as depth increases), 0.5, 1 (huge impact as depth increases).
-   *
-   * Warning: First arbitrary will not be used if its weight is set to zero.
-   *
-   * @remarks Since 2.14.0
-   */
   depthFactor?: number;
-  /**
-   * Maximal authorized depth.
-   * Once this depth has been reached only the first arbitrary will be used.
-   *
-   * Warning: Contrary to others, first arbitrary will be used even if its weight is set to zero.
-   *
-   * @remarks Since 2.14.0
-   */
   maxDepth?: number;
-  /**
-   * Depth identifier can be used to share the current depth between several instances.
-   *
-   * By default, if not specified, each instance of frequency will have its own depth.
-   * In other words: you can have depth=1 in one while you have depth=100 in another one.
-   *
-   * @remarks Since 2.14.0
-   */
   depthIdentifier?: string;
 };
 
-/**
- * @internal
- */
-function isFrequencyContraints(
-  param: FrequencyContraints | WeightedArbitrary<unknown> | undefined
-): param is FrequencyContraints {
-  return param != null && typeof param === 'object' && !('arbitrary' in param);
-}
-
-/**
- * For one of the values generated by `...warbs` - the probability of selecting the ith warb is of `warb[i].weight / sum(warb[j].weight)`
- *
- * **WARNING**: It expects at least one (Arbitrary, weight)
- *
- * @param warbs - (Arbitrary, weight)s that might be called to produce a value
- *
- * @remarks Since 0.0.7
- * @public
- */
-function frequency<Ts extends WeightedArbitrary<unknown>[]>(...warbs: Ts): Arbitrary<FrequencyValue<Ts>>;
-/**
- * For one of the values generated by `...warbs` - the probability of selecting the ith warb is of `warb[i].weight / sum(warb[j].weight)`
- *
- * **WARNING**: It expects at least one (Arbitrary, weight)
- *
- * @param constraints - Constraints to be applied when generating the values
- * @param warbs - (Arbitrary, weight)s that might be called to produce a value
- *
- * @remarks Since 0.0.7
- * @public
- */
-function frequency<Ts extends WeightedArbitrary<unknown>[]>(
-  constraints: FrequencyContraints,
-  ...warbs: Ts
-): Arbitrary<FrequencyValue<Ts>>;
-function frequency<Ts extends WeightedArbitrary<unknown>[]>(
-  ...args: [...Ts] | [FrequencyContraints, ...Ts]
-): Arbitrary<FrequencyValue<Ts>> {
-  // TODO With TypeScript 4.0 it will be possible to properly define typings for `frequency(...arbs, constraints)`
-  const label = 'fc.frequency';
-  const constraints = args[0];
-  if (isFrequencyContraints(constraints)) {
-    return FrequencyArbitrary.from(args.slice(1) as WeightedArbitrary<FrequencyValue<Ts>>[], constraints, label);
-  }
-
-  return FrequencyArbitrary.from(args as WeightedArbitrary<FrequencyValue<Ts>>[], {}, label);
-}
-
-export { frequency };
+/** @internal */
+type _FrequencyArbitraryContext<T> = {
+  selectedIndex: number;
+  originalBias: number | undefined;
+  originalContext: unknown;
+  clonedMrngForFallbackFirst: Random | null;
+  cachedGeneratedForFirst?: NextValue<T>;
+};
