+++ src/arbitrary/_internals/helpers/DepthContext.ts
