+++ src/check/arbitrary/WebArbitrary.ts
@@ -5,8 +5,8 @@ import { buildAlphaNumericPercentArb } from './helpers/SpecificCharacterRange';
 import { domain, hostUserInfo } from './HostArbitrary';
 import { nat } from '../../arbitrary/nat';
 import { ipV4, ipV4Extended, ipV6 } from './IpArbitrary';
-import { oneof } from './OneOfArbitrary';
-import { option } from './OptionArbitrary';
+import { oneof } from '../../arbitrary/oneof';
+import { option } from '../../arbitrary/option';
 import { stringOf } from './StringArbitrary';
 import { tuple } from './TupleArbitrary';
 import { Arbitrary } from './definition/Arbitrary';
