+++ test/unit/arbitrary/OptionArbitrary.spec.ts
@@ -1,11 +1,11 @@
 import * as prand from 'pure-rand';
-import * as fc from '../../../../lib/fast-check';
+import * as fc from '../../../lib/fast-check';
 
-import { option } from '../../../../src/check/arbitrary/OptionArbitrary';
-import { Random } from '../../../../src/random/generator/Random';
+import { option } from '../../../src/arbitrary/option';
+import { Random } from '../../../src/random/generator/Random';
 
-import * as stubArb from '../../stubs/arbitraries';
-import * as stubRng from '../../stubs/generators';
+import * as stubArb from '../stubs/arbitraries';
+import * as stubRng from '../stubs/generators';
 
 describe('OptionArbitrary', () => {
   describe('option', () => {
